fix(prevoyance): trois methodes de provisionnement, trois resultats
D08 -- `int()` tronquait vers zero une ESPERANCE de nombre d invalides.
Mesure : PM Rentes IP = 0 EUR jusqu a 496 assures, puis 482 895 EUR au 497e.
Un assure de plus faisait apparaitre un demi-million d euros de provision.
L asymetrie avec `_psap_ip`, qui plancher a 1 sur la MEME quantite, montrait
qu un des deux etait faux. Apres : 482 864 EUR a 496, 483 837 EUR a 497 --
continue, et croissante.

D09 -- sans primes exogenes, l a priori du Bornhuetter-Ferguson valait
`ult_cl x lr`, c est-a-dire l ultime projete par Chain Ladder de l annee
meme qu on cherche a estimer. Le BF degenerait en `lr x CL`, identiquement.
L a priori se construit desormais sur le VOLUME OBSERVE des annees matures,
qui ne depend pas de l annee estimee.

  BF / CL, sur trois CADENCES differentes (et non trois homotheties, qui
  multiplient CL et BF dans le meme rapport et ne prouvent rien) :
    avant      0,680000   0,680000   0,680000
    apres      0,952010   0,971043   0,586831

LE CAS MACK EST DIFFERENT, ET MON AUDIT LE QUALIFIAIT MAL. `Mack - CL =
0,000000` est mathematiquement NORMAL : la reserve centrale de Mack (1993)
EST celle du Chain Ladder ; Mack n apporte pas un second point, il apporte
la VARIANCE autour du premier. Le defaut n est donc pas l egalite -- c est
d avoir compte Mack comme une estimation INDEPENDANTE dans un
"CV inter-methodes" de 16,89 %, presente comme une incertitude de modele.
Le CV ne porte plus que sur des methodes reellement distinctes, et il PUBLIE
sa base et ses exclusions : 16,89 % -> 1,44 % sur
['bornhuetter_ferguson', 'chain_ladder'], exclues ['mack_1993'].

D10 -- `ra = 0,35 x BE x 0,06` vaut 2,10 % du BE, strictement inferieur au
plancher de 3,00 % : le `max` retournait le plancher pour TOUTE valeur
de BE, le cout du capital n avait aucun effet, et l etiquette publiee
annoncait "CoC 6 %". Balayage 10 kEUR -> 100 MEUR : RA/BE restait a 3,00 %.
Le calcul est conserve -- le plancher mord effectivement -- mais l agent DIT
desormais lequel des deux termes l emporte, et l etiquette suit la valeur.

D11 -- AU TEXTE DU REGLEMENT, et ce n etait donc pas un arbitrage.
Le sigma de PRIMES etait applique au Best Estimate. Le RD (UE) 2015/35
definit DEUX mesures de volume distinctes :
  V(prem,s) = max(Ps ; P(last,s)) + FP(existing,s) + FP(future,s)
              -- art. 147 par. 3, ou Ps et P(last,s) sont des PRIMES ;
  V(res,s)  = meilleure estimation des sinistres a payer -- art. 147 par. 6.
Le BE etait donc compte DEUX FOIS, avec deux sigmas, et l exposition aux
primes etait totalement absente : un portefeuille doublant ses primes sans
changer ses provisions voyait son SCR inchange.

Les sigmas eux-memes ne venaient d aucun segment. Annexe XIV, segment 2
"assurance de protection du revenu" (lignes d activite 2 et 14) :
sigma primes = 8,5 %, sigma reserves = 14 %. Le code utilisait 10 % et 11 %.
Le 11 % coincide EXACTEMENT avec le sigma reserves du segment 3,
"indemnisation des travailleurs" -- que ce soit la cause reste une
HYPOTHESE non instruite ; le fait etabli est que ces valeurs ne sont pas
celles de l activite.

`test_p3_provisionnement.py::test_t6` RECALCULAIT la troncature `int()` pour
en faire son attendu : il verrouillait le defaut qu il aurait du denoncer.
Son attendu devient l esperance continue, DANS CE LOT.

Le sceau `test_sceau_provisions.py` a ete eprouve par trois plants :
troncature retablie -> 2 tests rougissent ; a priori BF redevenu l ultime CL
-> 1 test ; SCR de prime ramene sur le BE -> 1 test.

  pytest             185 passed -> 195 passed (+ 14 subtests)
  unittest discover  185 OK     -> 195 OK

+++ direction_sante_prevoyance/prevoyance/p3_provisionnement/test_p3_provisionnement.py
@@ -486,7 +486,12 @@ def _impl_test_t6_provisions_long_terme():
     rente_an = sal * rente                     # 27 000 €
     v        = 1.0 / (1 + taux_a)
     annuite  = sum(v**t for t in range(1, int(dur) + 1))  # ~18.9 ans
-    nb_inv   = max(0, int(nb * tip * 0.60))    # = int(500 × 0.00336 × 0.60) = 1
+    # ⚠️ ATTENDU CORRIGÉ LE 12/09/2026, en même temps que le contrat.
+    # Ce test RECALCULAIT la troncature `int()` pour en faire son attendu : il
+    # verrouillait donc le défaut qu'il aurait dû dénoncer. Cette troncature
+    # faisait valoir la PM 0 EUR jusqu'à 496 assurés, puis 482 895 EUR au 497e.
+    # Une espérance de nombre d'invalides est continue.
+    nb_inv   = max(0.0, nb * tip * 0.60)       # 500 × 0.00336 × 0.60 = 1.008
     pm_att   = nb_inv * rente_an * annuite
 
     pm_obt = r["pm_rentes_ip"]
